Hive Scene foundation
Added HexGridManager and HexTile scripts, a few objects in hive scene, and a hex tile asset. HexTile is basically the same as Tile, but has an unused untested method to change the sprite.
The plan is to control the hex tiles from the manager through an indexed list and several methods to alter the state of the tiles.